// src/constants/enums/messages.ts
 export enum JOB_SUCCESS_MESSAGES {
  JOB_CREATED = "Job created successfully",
  JOB_UPDATED = "Job updated successfully",
  JOB_DELETED = "Job deleted successfully",
  JOB_FETCHED = "Job fetched successfully",
  CREATED_CANDIDATES_ADDED_TO_JOB = "Created candidates added to job successfully",
  CANDIDATES_FETCHED = "Candidates fetched successfully",
}
export enum JOB_ERROR_MESSAGES {  
  JOB_NOT_FOUND = "Job not found",
}